Evals: Move score calculation to ScoreEvals
